Support updating of kits with checked out branches
The following assumption is made:
Either a kit git submodule will be in a detached HEAD state, or it will
have a specific branch checked out (these are the two common use cases).
Based on that assumption, in the former case the kit's git submodule
shall be updated via `git submodule update [...]` to the commit
recorded in meta-repo for it, in the latter the respective branch
it's on shall be updated via `git pull [...]`.