Update repo_can_write_test() to not use os.makedirs().
This was creating a meta-repo directory if it didn't
already exist, preventing a new meta-repo from being
cloned.

The rewrite of this method now just recursively (loops)
trying to write a .touch file to see if it has
permissions somewhere to write in the /var/git/meta-
repo path, or any parent dir that exists.